[Repair] FoodFighter 동기화 작업 복구 & DB Query 수정
- FoodPrefabs에 PhotonView 추가.
- Food와 FoodInteraction Scripts 복구 작업.
- FoodInteraction에서 Player의 Scale을 변경 시키는 방식을 PhotonView -> RPC로 변경.
- UpdateRelationship 의 Query Syntax 수정.
