+++ Assets/_ImportedAssets/FoodFighter/Bread_Dish.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 5440368780815714985}
   - component: {fileID: 3943852122564286694}
   - component: {fileID: 4975722285801852554}
+  - component: {fileID: 3705429423598584748}
   m_Layer: 10
   m_Name: Bread_Dish
   m_TagString: Untagged
@@ -62,6 +63,29 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.3, y: 0.25, z: 0.65}
   m_Center: {x: 0, y: 0.15, z: 0}
+--- !u!114 &3705429423598584748
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5923320022609438542}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents:
+  - {fileID: 3943852122564286694}
+  sceneViewId: 0
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1001 &7671534767632150924
 PrefabInstance:
   m_ObjectHideFlags: 0
