+++ Assets/_ImportedAssets/FoodFighter/Fish_Cooked_Dish.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 7265380920706198346}
   - component: {fileID: 2980794612660346574}
   - component: {fileID: 4287589612662951136}
+  - component: {fileID: 8723205471428476207}
   m_Layer: 10
   m_Name: Fish_Cooked_Dish
   m_TagString: Untagged
@@ -62,6 +63,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 1, y: 0.06, z: 1}
   m_Center: {x: 0, y: 0.1, z: 0}
+--- !u!114 &8723205471428476207
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5676414703864687021}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents: []
+  sceneViewId: 0
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1001 &2309635906173342212
 PrefabInstance:
   m_ObjectHideFlags: 0
