+++ Assets/_ImportedAssets/FoodFighter/Pie_Dish_01.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 4884807718529108578}
   - component: {fileID: 4670942791954478846}
   - component: {fileID: 5843053045163150454}
+  - component: {fileID: 7256040231230720299}
   m_Layer: 10
   m_Name: Pie_Dish_01
   m_TagString: Untagged
@@ -62,6 +63,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.3, y: 0.1, z: 0.3}
   m_Center: {x: 0, y: 0.05, z: 0}
+--- !u!114 &7256040231230720299
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5791893569924701036}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents: []
+  sceneViewId: 0
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1001 &4399687651241241995
 PrefabInstance:
   m_ObjectHideFlags: 0
