+++ Assets/_ImportedAssets/FoodFighter/Ribs_Dish.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 5495978030903379952}
   - component: {fileID: 3688977490257563687}
   - component: {fileID: 7300306932463121821}
+  - component: {fileID: 5863177960271021967}
   m_Layer: 10
   m_Name: Ribs_Dish
   m_TagString: Untagged
@@ -62,6 +63,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.3, y: 0.1, z: 0.3}
   m_Center: {x: 0, y: 0.05, z: 0}
+--- !u!114 &5863177960271021967
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7712684921540237696}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents: []
+  sceneViewId: 0
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1001 &3010091627467961123
 PrefabInstance:
   m_ObjectHideFlags: 0
