+++ Assets/_ImportedAssets/FoodFighter/Turkey_Cooked_Dish.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 7056397504284333942}
   - component: {fileID: 5302928899630654287}
   - component: {fileID: 56737721278157148}
+  - component: {fileID: -1021068863046230190}
   m_Layer: 10
   m_Name: Turkey_Cooked_Dish
   m_TagString: Food
@@ -62,6 +63,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.7, y: 0.2, z: 0.7}
   m_Center: {x: 0, y: 0.12, z: 0}
+--- !u!114 &-1021068863046230190
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5294337276658325998}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: aa584fbee541324448dd18d8409c7a41, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  ObservedComponentsFoldoutOpen: 1
+  Group: 0
+  prefixField: -1
+  Synchronization: 3
+  OwnershipTransfer: 0
+  observableSearch: 2
+  ObservedComponents: []
+  sceneViewId: 0
+  InstantiationId: 0
+  isRuntimeInstantiated: 0
 --- !u!1001 &2596103726973502625
 PrefabInstance:
   m_ObjectHideFlags: 0
