+++ Assets/_ImportedAssets/Photon/PhotonUnityNetworking/Resources/PhotonServerSettings.asset
@@ -44,6 +44,9 @@ MonoBehaviour:
   - HeresWhy
   - StartParticle
   - GetMagic
+  - CheckFoodState
+  - EatFoodState
+  - RegenerateFoodState
   DisableAutoOpenWizard: 1
   ShowSettings: 1
   DevRegionSetOnce: 0
