--- Assets/_Scripts/Keyboard/Lobby1_Interact/FoodFighter/Food.cs
@@ -1,45 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.Events;
-
-public enum EFoodSatietyLevel
-{
-    None,
-    Small,
-    Big
-};
-
-
-
-public class Food : InteracterableObject
-{
-    public static UnityEvent<EFoodSatietyLevel> OnEated = new UnityEvent<EFoodSatietyLevel>();
-
-    [SerializeField] EFoodSatietyLevel _foodSatietyLevel;
-    [SerializeField] GameObject _food;
-    private static readonly YieldInstruction _waitSecondRegenerate = new WaitForSeconds(60f);
-
-
-    public override void Interact()
-    {
-        base.Interact();
-
-        OnEated.Invoke(_foodSatietyLevel);
-
-        _food.SetActive(false);
-
-        StartCoroutine(RegenerateFood());
-    }
-
-    IEnumerator RegenerateFood()
-    {
-        yield return _waitSecondRegenerate;
-
-        _food.SetActive(true);
-
-        yield return null;
-    }
-
-
-}
