+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,7 +9,7 @@ EditorBuildSettings:
     path: Assets/_Scenes/BuildScene/Login.unity
     guid: 6f6fb3e7831651c47a96f3bb722cd4dd
   - enabled: 1
-    path: Assets/Scenes/StartRoom.unity
+    path: Assets/_Scenes/BuildScene/Lobbys/StartRoom.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
   - enabled: 0
     path: Assets/_Scenes/BuildScene/Lobbys/StartRoom.unity
